add min duration
