+++ DOCKER-CONFIG.md
@@ -11,7 +11,7 @@ docker run -p 8080:8080 -e GOOGLE_API_KEY=<api key here> -v /<audio download pat
 
 Docker run command (all parameters)
 ```
-docker run -p 8080:8080 -e GOOGLE_API_KEY=<api key here> -v /<audio download path here>:/config -e TRUSTED_HOSTS=<add hosts here> -e TOKEN=<add secure token here> -e CRON="0 0 * * 0" -e SPONSORBLOCK_CATEGORIES="sponsor" -e COOKIES_FILE <cookies file path here> ikoyhn/go-podcast-sponsor-block
+docker run -p 8080:8080 -e GOOGLE_API_KEY=<api key here> -v /<audio download path here>:/config -e TRUSTED_HOSTS=<add hosts here> -e TOKEN=<add secure token here> -e CRON="0 0 * * 0" -e SPONSORBLOCK_CATEGORIES="sponsor" -e COOKIES_FILE <cookies file path here> -e MIN_DURATION=<min duration of video to grab> ikoyhn/go-podcast-sponsor-block
 ```
 
 ## Docker Compose Templates
@@ -63,3 +63,4 @@ volumes:
 | `-e CRON` | By default a cron job will be run weekly to delete any podcast episode files that havent been access in over a week, if you want to modify when this runs you can set the cron here ([CRON examples](https://crontab.guru/))| No |
 | `-e SPONSORBLOCK_CATEGORIES` | Customize the categories that you would like to remove from your podcasts. String separated by `,` with possible values `sponsor,selfpromo,interaction,intro,outro,preview,music_offtopic,filler`. Default: `sponsor` | No |
 | `-e COOKIES_FILE` | Run the app once for the config folder to be created then store your cookies folder in the root of the config folder and set the filename for the docker var. Set this if you want to use custom cookies for YT-DLP| No |
+| `-e MIN_DURATION` | To filter out YT shorts there is a minimum duration a video has to be in order to grab it. The default is 5min, modify as needed. | No |
